requirements.txt
streamlit
google-generativeai
python-dotenv
Flask
openai